Fix: 여행 세부계획 응답 dto 분리
# Conflicts:
#	src/main/java/org/solcation/solcation_be/domain/travel/TravelController.java
