+++ src/main/java/org/solcation/solcation_be/domain/travel/dto/MoveDayReq.java
@@ -1,0 +1,10 @@
+package org.solcation.solcation_be.domain.travel.dto;
+
+public record MoveDayReq(
+        String crdtId,
+        int newDay,
+        String prevCrdtId,
+        String nextCrdtId,
+        String clientId,
+        Long opTs
+) {}
