+++ src/main/java/org/solcation/solcation_be/domain/travel/dto/MoveWithinReq.java
@@ -1,0 +1,9 @@
+package org.solcation.solcation_be.domain.travel.dto;
+
+public record MoveWithinReq(
+        String crdtId,
+        String prevCrdtId,
+        String nextCrdtId,
+        String clientId,
+        Long opTs
+) {}
