+++ src/main/java/org/solcation/solcation_be/domain/travel/dto/UpdateReq.java
@@ -1,0 +1,11 @@
+package org.solcation.solcation_be.domain.travel.dto;
+
+public record UpdateReq(
+        String crdtId,
+        String pdPlace,
+        String pdAddress,
+        int pdCost,
+        Long tcPk,
+        String clientId,
+        Long opTs
+) {}
